% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_methods.R
\name{ConsensusSignatures}
\alias{ConsensusSignatures}
\title{Consensus Signatures}
\usage{
ConsensusSignatures(object, bind = TRUE)
}
\arguments{
\item{object}{A class "ScudoResults" object.}

\item{bind}{Logic, default = TRUE, returns concatenated or separated
Consensus Signatures.}
}
\description{
Returns Consensus Signatures. If bind = TRUE (default)
returns single, else returns a list composed of separated "UpSignatures" and
"DownSignatures".
}
