% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDoc.R
\docType{package}
\name{rScudo-package}
\alias{rScudo-package}
\title{test1}
\description{
test2
}
\details{
test3
}
\examples{
example("scudoTrain")

}
\seealso{
\code{\link{scudoTrain}}, \code{\link{scudoNetwork}},
\code{\link{scudoClassify}}
}
\author{
Matteo Ciciani \email{matteo.ciciani@gmail.com}, Thomas Cantore
\email{cantorethomas@gmail.com}
}
