% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudo.R
\name{scudo}
\alias{scudo}
\title{Performs Scudo functions on Expression Data}
\usage{
scudo(expressionData, groups, nTop, nBottom, pValue, prepro = TRUE,
  featureSel = TRUE, p.adj = "none")
}
\arguments{
\item{expressionData}{Data frame object that contains expressionData with
different samples as columns.}

\item{groups}{Factor containing groups labels for samples in
expressionData.}

\item{nTop}{Number of UpSignatures to be selected.}

\item{nBottom}{Number of DownSignatures to be selected.}

\item{pValue}{pValue for optional feature selection. If no feature selection
is performed, pValue is not used.}

\item{prepro}{Preprocessing on expressionData performed through a
Normalization step. Default = \code{TRUE}: if \code{FALSE}
no normalization is performed.}

\item{featureSel}{Feature selection performed through Wilcoxon-Mann-Withney
test, or with Kruskal-Wallis test. Default = \code{TRUE}: if \code{FALSE}
no feature selection is performed.}

\item{p.adj}{p.adj optionally performed on feature selection step. Default
\code{p.adj = "none"}. Look at \code{\link[stats]{p.adjust.methods}} for the
possible adjustments.}
}
\value{
S4 class object \linkS4class{ScudoResults}.
}
\description{
Performs optional normalization and feature selection, then creates a
Scudo Results object.
}
