% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudoClassify.R
\name{scudoClassify}
\alias{scudoClassify}
\title{Performes classification using SCUDO}
\usage{
scudoClassify(trainExpData, testExpData, N, nTop, nBottom, trainGroups,
  maxDist = 1, weighted = TRUE, complete = FALSE, beta = 1,
  alpha = 0.1, norm = TRUE, featureSel = TRUE, parametric = FALSE,
  pAdj = "none", distFun = NULL)
}
\arguments{
\item{trainExpData}{either an \code{\link[Biobase]{ExpressionSet}},
a data.frame or a matrix of gene expression data, with a column for
each sample and a row for each feature, to be used as a training set}

\item{testExpData}{either an \code{\link[Biobase]{ExpressionSet}},
a data.frame or a matrix of gene expression data, with a column for
each sample and a row for each feature, to be used as a testing set}

\item{N}{a number between 0 and 1, representing the fraction of the
signature-to-signature distances that will be used to draw the graph}

\item{nTop}{number of up-regulated features to include in the signatures}

\item{nBottom}{number of down-regulated features to include in the
signatures}

\item{trainGroups}{factor containing group labels for each sample in
\code{trainExpData}}

\item{maxDist}{intger. Only nodes with a distance from a testing node less or
equal to \code{maxDist} are used to perform the classification.}

\item{weighted}{logical, whether to consider the distances associated to the
edges to compute the scores for the classification. For a description of the
classification method, see Details below}

\item{complete}{logical, whether to consider all the nodes in the training
set to perform the classification. If TRUE, the arguments \code{N},
\code{maxDist} and \code{weighted} are ignored. For a description of the
classification method, see Details below}

\item{beta}{a coefficient used to down-weight the influence of distant nodes
on the classification outcome. For a description of the
classification method, see Details below}

\item{alpha}{p-value cutoff for the optional feature selection step. If
feature selection is skipped, alpha is ignored}

\item{norm}{logical, whether or not to normalize the expression data. See
Details for a description of the normalization used}

\item{featureSel}{logical, whether or not to perform a feature selection.
Feature selection is performed using one of four tests: Student's t-test,
ANOVA, Wilcoxon-Mann-Withney test, or Kruskal-Wallis test. The test
used depends on the number of groups and the \code{parametric} argument}

\item{parametric}{logical, whether to use a parametric or a non-parametric
test for the feature selection}

\item{pAdj}{pAdj method to use to adjust the p-values in the feature
selection step. See \code{\link[stats]{p.adjust.methods}} for a list of
adjustment methods}

\item{distFun}{the function used to compute the distance between two
samples. See Details for the specification of this function}
}
\value{
A \code{data.frame} containing classification scores for each sample
in \code{testExpData}
}
\description{
Performes classification using SCUDO
}
\details{
Placeholder
}
