% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScudoCytoscape.R
\name{ScudoCytoscape}
\alias{ScudoCytoscape}
\title{Create a Cytoscape network from the output of ScudoNetwork}
\usage{
ScudoCytoscape(igraph, title = NULL, collection = "SCUDO",
  base.url = NULL)
}
\arguments{
\item{igraph}{object of class igraph, like the result of
\code{\link{ScudoNetwork}}.}

\item{title}{The title of the network. If NULL it defaults to
\code{deparse(substitute(igraph))}.}

\item{collection}{The name of the collection.}

\item{base.url}{See \code{\link[RCy3]{createNetworkFromIgraph}}}
}
\value{
The network SUID (an integer).
}
\description{
Visualize results of scudo analysis in Cytoscape
}
