% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudoNetwork.R
\docType{methods}
\name{scudoNetwork}
\alias{scudoNetwork}
\alias{scudoNetwork,scudoResults-method}
\title{Create graph from a scudoResults object}
\usage{
scudoNetwork(object, N, colors = character())
}
\arguments{
\item{object}{A \code{\link{scudoResults}} object}

\item{N}{A number between 0 and 1. The fraction of all of the
signature-to-signature distances that will be used for drawing the graph.}

\item{colors}{A character vector}
}
\description{
A function to create an \code{\link[igraph:igraph-package]{igraph}} graph
from a \code{\link{scudoResults}} object.
}
\details{
The distance matrix in the scudoResults
}
