% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/accessors.R
\docType{class}
\name{ScudoResults-class}
\alias{ScudoResults-class}
\alias{ScudoResults}
\alias{DistMatrix}
\alias{DistMatrix,ScudoResults-method}
\alias{DistMatrix,ScudoResults-method}
\title{Class "ScudoResults"}
\description{
This is a class that represents the output of the functions
-- link to scudo functions --.
}
\details{
This class provides a structure to represent the results of the scudo
functions. It contains a distance matrix, i.e. a non-negative symmetric
matrix.
}
\section{Slots}{

\describe{
\item{\code{DistMatrix:}}{Object of class "matrix".}

\item{\code{UpSignatures:}}{Object of class "data.frame".}

\item{\code{DownSignatures:}}{Object of class "data.frame."}

\item{\code{Groups:}}{Object of class "factor".}

\item{\code{ConsensusUpSignatures:}}{Object of class "data.frame".}

\item{\code{ConsensusDownSignatures:}}{Object of class "data.frame".}

\item{\code{SelectedFeatures:}}{Object of class "character".}

\item{\code{Params:}}{Object of class "list".}
}}

\section{Methods}{

\describe{
  \item{\code{DistMatrix}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the distance matrix.}
  \item{\code{UpSignatures}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the .}
  \item{\code{DownSignatures}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the .}
  \item{\code{Groups}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the .}
  \item{\code{ConsensusUpSignatures}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the .}
  \item{\code{ConsensusDownSignatures}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the .}
  \item{\code{SelectedFeatures}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the .}
  \item{\code{Params}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the .}
}
}

