% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_methods.R
\name{ConsensusSignatures}
\alias{ConsensusSignatures}
\title{Consensus Signatures}
\usage{
\code{ConsensusSignatures(object, ...)}
}
\arguments{
\item{object}{A "ScudoResults" class object}

\item{bind}{logical, indicating wether or not concatenating Up and Down
Singatures}
}
\description{
If bind = TRUE (default) returns a list containing all Consensus
Signatures, else returns a list divided into up and down.
}
\examples{
## With 'bind = TRUE' for a generic ScudoResults object

ConsensusSignatures(ScudoResObj, bind = TRUE)

[1] "l" "m" "n" "o" "p" "q" "r" "s" "t" "u" "v" "w" "x" "y"

## 'bind = FALSE'

$UpSignatures
[1] "l" "m" "n" "o" "p" "q" "r"

$DownSignatures
[1] "s" "t" "u" "v" "w" "x" "y"
}
