% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_methods.R
\name{ConsensusSignatures}
\alias{ConsensusSignatures}
\title{Consensus Signatures}
\usage{
ConsensusSignatures(object, bind = TRUE)
}
\arguments{
\item{object}{A "ScudoResults" class object}

\item{bind}{logical, indicating wether or not concatenating Up and Down
Singatures}
}
\description{
If bind = TRUE (default) returns a list containing all Consensus
Signatures, else returns a list divided into up and down.
}
