% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudo.R
\name{scudo}
\alias{scudo}
\title{Performs SCUDO on Expression Data}
\usage{
scudo(expressionData, groups, nTop, nBottom, pValue = 0.1,
  prepro = TRUE, featureSel = TRUE, pAdj = "none")
}
\arguments{
\item{expressionData}{data.frame object that contains expressionData with
different samples as columns.}

\item{groups}{factor containing groups labels for samples in expressionData.}

\item{nTop}{number of up-regulated features to include in the signatures.}

\item{nBottom}{number of down-regulated features to include in the signatures.}

\item{pValue}{pValue cutoff for optional feature selection. If no feature
selection is performed, pValue is ignored.}

\item{prepro}{logical, whether or not to normalize the expression data. See
Details for a description of the normalization used.}

\item{featureSel}{Logical, whether or not to perform feature selection.
Feature selection performed through a Wilcoxon-Mann-Withney test, or with a
Kruskal-Wallis test, depending on the number of groups.}

\item{pAdj}{pAdj optionally performed on feature selection step. Default
\code{pAdj = "none"}. Look at \code{\link[stats]{p.adjust.methods}} for
the possible adjustments.}
}
\value{
S4 class object \linkS4class{scudoResults}.
}
\description{
Performs optional normalization and feature selection, then creates a Scudo
Results object.
}
\details{
\code{scudo} performs a normalization based on mean features expression
levels in different groups, in order to increase the sensitivity for the
subsequent analysis.

The feature selection process is helpful for subsequent analysis. It is also
possible to compute optionally a pValue adjustment for stricter feature
selection.
}
