% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudoClassify.R
\name{scudoClassify}
\alias{scudoClassify}
\title{Performes classification using SCUDO}
\usage{
scudoClassify(trainExpData, testExpData, N, nTop, nBottom,
    trainGroups, maxDist = 1, weighted = TRUE, complete = FALSE, beta = 1,
    alpha = 0.1, foldChange = TRUE, featureSel = TRUE, parametric = FALSE,
    pAdj = "none", distFun = NULL)
}
\arguments{
\item{trainExpData}{either an \code{\link[Biobase]{ExpressionSet}},
a data.frame or a matrix of gene expression data, with a column for
each sample and a row for each feature, to be used as a training set}

\item{testExpData}{either an \code{\link[Biobase]{ExpressionSet}},
a data.frame or a matrix of gene expression data, with a column for
each sample and a row for each feature, to be used as a testing set}

\item{N}{a number between 0 and 1, representing the fraction of the
signature-to-signature distances that will be used to draw the graph}

\item{nTop}{number of up-regulated features to include in the signatures}

\item{nBottom}{number of down-regulated features to include in the
signatures}

\item{trainGroups}{factor containing group labels for each sample in
\code{trainExpData}}

\item{maxDist}{intger. Only nodes with a distance from a testing node less or
equal to \code{maxDist} are used to perform the classification.}

\item{weighted}{logical, whether to consider the distances associated to the
edges to compute the scores for the classification. For a description of the
classification method, see Details below}

\item{complete}{logical, whether to consider all the nodes in the training
set to perform the classification. If TRUE, the arguments \code{N},
\code{maxDist} and \code{weighted} are ignored. For a description of the
classification method, see Details below}

\item{beta}{a coefficient used to down-weight the influence of distant nodes
on the classification outcome. For a description of the
classification method, see Details below}

\item{alpha}{p-value cutoff for the optional feature selection step. If
feature selection is skipped, alpha is ignored}

\item{foldChange}{logical, whether or not to compute fold-changes from
expression data}

\item{featureSel}{logical, whether or not to perform a feature selection.
Feature selection is performed using one of four tests: Student's t-test,
ANOVA, Wilcoxon-Mann-Withney test, or Kruskal-Wallis test. The test
used depends on the number of groups and the \code{parametric} argument}

\item{parametric}{logical, whether to use a parametric or a non-parametric
test for the feature selection}

\item{pAdj}{pAdj method to use to adjust the p-values in the feature
selection step. See \code{\link[stats]{p.adjust.methods}} for a list of
adjustment methods}

\item{distFun}{the function used to compute the distance between two
samples. See Details for the specification of this function}
}
\value{
A \code{list} containing the predictions for each sample in
\code{testExpData} and the classification scores used to generate the
predictions.
}
\description{
placeholder
}
\details{
Placeholder
}
\examples{
expData <- data.frame(a = 1:10, b = 2:11, c = 10:1, d = 11:2,
    e = c(1:4, 10:5), f = c(7:10, 6:1), g = c(8:4, 1:3, 10, 9),
    h = c(6:10, 5:1), i = c(5:1, 6:10))
rownames(expData) <- letters[1:10]
groups <- factor(c(1,1,1,2,2,2,1,1,1))
inTrain <- 1:5

# perform classification
res <- scudoClassify(expData[, inTrain], expData[, -inTrain], 0.9, 3, 3,
    groups[inTrain], featureSel = FALSE)

#explore predictions
predictions <- res$predicted
scores <- res$scores

}
\seealso{
\code{\link{scudoTrain}}
}
\author{
Matteo Ciciani \email{matteo.ciciani@gmail.com}
}
