% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudoClassify.R
\name{scudoClassify}
\alias{scudoClassify}
\title{Performes classification using SCUDO}
\usage{
scudoClassify(trainExpData, testExpData, N, nTop, nBottom, trainGroups,
  testGroups = NULL, alpha = 0.1, norm = TRUE, groupedNorm = FALSE,
  featureSel = TRUE, parametric = FALSE, pAdj = "none",
  distFun = NULL, neighbours = 1, weighted = TRUE, pruned = FALSE)
}
\description{
Placeholder
}
