% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScudoCytoscape.R
\name{ScudoCytoscape}
\alias{ScudoCytoscape}
\title{ScudoCytoscape}
\usage{
ScudoCytoscape(scudoIgraph, title = NULL, collection = "SCUDO",
  base.url = NULL)
}
\arguments{
\item{scudoIgraph}{object of class igraph, as the result of
\code{\link{ScudoNetwork}}.}

\item{title}{The title of the network. If NULL it defaults to
\code{deparse(substitute(scudoIgraph))}.}

\item{collection}{The name of the collection.}

\item{See}{\code{\link[RCy3]{createNetworkFromIgraph}}}
}
\value{
network SUID
}
\description{
Visualize results of scudo analysis in Cytoscape
}
