% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScudoNetwork.R
\docType{methods}
\name{ScudoNetwork}
\alias{ScudoNetwork}
\alias{ScudoNetwork,ScudoResults-method}
\title{ScudoNetwork}
\usage{
ScudoNetwork(object, N, colors = character())
}
\arguments{
\item{object}{A ScudoResults object}

\item{N}{A numeric value.}

\item{colors}{A character vector}
}
\description{
A function to convert a ScudoResult in an igraph object
}
\details{
Details about the function
}
