% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudoNetwork.R
\docType{methods}
\name{scudoNetwork}
\alias{scudoNetwork}
\alias{scudoNetwork,scudoResults-method}
\title{scudoNetwork}
\usage{
scudoNetwork(object, N, colors = character())
}
\arguments{
\item{object}{A scudoResults object}

\item{N}{A numeric value.}

\item{colors}{A character vector}
}
\description{
A function to convert a ScudoResult in an igraph object
}
\details{
Details about the function
}
