% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudoPredict.R
\name{scudoPredict}
\alias{scudoPredict}
\title{Performs scudoPredict on test Expression Data}
\usage{
scudoPredict(trainScudoRes, testExpData, testGroups, nTop, nBottom,
  prepro = TRUE, groupedNorm = TRUE, distFun = NULL)
}
\arguments{
\item{trainScudoRes}{scudoResults object used as training object.}

\item{testExpData}{data.frame object containing expressionData used as test.}

\item{testGroups}{factor containing groups labels for samples in testExpData.}

\item{nTop}{number of up-regulated features to include in the signatures.}

\item{nBottom}{number of down-regulated features to include in the signatures.}

\item{prepro}{logical, whether or not to normalize the test expression data.
See Details for a description of the normalization used.}

\item{groupedNorm}{logical, whether or not to performed grouped
normalization. See Details for a description of the normalization used}

\item{distFun}{the function used to compute the distance between two
samples. See Details}
}
\value{
S4 class object \linkS4class{scudoResults}.
}
\description{
Performes SCUDO on test Expression Data using feature selected from a
previous computed train Scudo Results object.
}
\details{
\code{scudoPredict} works as a common predict function by testing on a
different expressionData object previous feature selection results obtained
in a train Scudo Result object. This could be helpful in order to check the
effectiveness of previously chosen parameters.
}
