% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/accessors.R
\docType{class}
\name{ScudoResults-class}
\alias{ScudoResults-class}
\alias{ScudoResults}
\alias{DistMatrix}
\alias{DistMatrix,ScudoResults-method}
\alias{DistMatrix,ScudoResults-method}
\title{Class "ScudoResults"}
\usage{
DistMatrix(object)
}
\description{
This is a class that represents the output of the functions
-- link to scudo functions --.
}
\details{
Details about the class
}
\section{Slots}{

\describe{
\item{\code{DistMatrix}}{Object of class "matrix".}

\item{\code{UpSignatures}}{Object of class "data.frame".}

\item{\code{DownSignatures}}{Object of class "data.frame."}

\item{\code{Groups}}{Object of class "factor".}

\item{\code{ConsensusUpSignatures}}{Object of class "data.frame".}

\item{\code{ConsensusDownSignatures}}{Object of class "data.frame".}

\item{\code{SelectedFeatures}}{Object of class "character".}

\item{\code{Params}}{Object of class "list".}
}}

\section{Methods}{

\describe{
  \item{One}{First item}
  \item{Two}{Second item}
}
}

