% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/accessors.R
\docType{class}
\name{ScudoResults-class}
\alias{ScudoResults-class}
\alias{ScudoResults}
\alias{DistMatrix}
\alias{DistMatrix,ScudoResults-method}
\alias{DistMatrix,ScudoResults-method}
\title{Class "ScudoResults"}
\description{
This is a class that represents the output of the functions -- link to scudo
functions --.
}
\details{
This class provides a structure to represent the results of the scudo
functions. It contains a distance matrix, i.e. a symmetric matrix with
non-negative numeric values.
}
\section{Slots}{

\describe{
\item{\code{DistMatrix}}{Object of class "matrix". A symmetric matrix with
non-negative numeric elements.}

\item{\code{UpSignatures}}{Object of class "data.frame". A data.frame with the same
colnames as DistMatrix, representing the up-regualted features in each
sample.}

\item{\code{DownSignatures}}{Object of class "data.frame". A data.frame with the same
colnames as DistMatrix, representing the down-regualted features in each
sample.}

\item{\code{Groups}}{Object of class "factor". It represents the groups used for the
normalization and the feature selection. It corresponds to the
\code{groups} argument in --link-- and --link--, but unused levels are
dropped.}

\item{\code{ConsensusUpSignatures}}{Object of class "data.frame". It contains the
consensus signatures of up-regulated features for each group.}

\item{\code{ConsensusDownSignatures}}{Object of class "data.frame".It contains the
consensus signatures of dowm-regulated features for each group.}

\item{\code{SelectedFeatures}}{Object of class "character". A vector of selected
features. If the feature selection was not performed, it contains every
feature present in the input of the scudo functions.}

\item{\code{Params}}{Object of class "list". A list of the parameters used to run
--link-- or --link--.}
}}

\section{Methods}{
 \describe{ \item{\code{DistMatrix}}{\code{signature(object
  = "ScudoResults")}: a method for obtaining the distance matrix.}
  \item{\code{UpSignatures}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the signature of up-regualted features in each
  sample.} \item{\code{DownSignatures}}{\code{signature(object =
  "ScudoResults")}: a method for obtaining the signature of down-regulated
  features in each sample.} \item{\code{Groups}}{\code{signature(object =
  "ScudoResults")}: a method for obtaining the groups used for normalization
  and feature selection.}
  \item{\code{ConsensusUpSignatures}}{\code{signature(object =
  "ScudoResults")}: a method for obtaining the consensus signatures of
  up-regualted features in each group.}
  \item{\code{ConsensusDownSignatures}}{\code{signature(object =
  "ScudoResults")}: a method for obtaining the consensus signatures of
  down-regulated features in each group.}
  \item{\code{SelectedFeatures}}{\code{signature(object = "ScudoResults")}: a
  method for obtaining the names of the features seleted. If no feature
  selection was performed, the names of every feature are returned.}
  \item{\code{Params}}{\code{signature(object = "ScudoResults")}: a method
  for obtaining the parameters that were used to generate the result.} }
}

