% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/accessors.R, R/show.R
\docType{class}
\name{scudoResults-class}
\alias{scudoResults-class}
\alias{scudoResults}
\alias{distMatrix}
\alias{distMatrix,scudoResults-method}
\alias{upSignatures}
\alias{upSignatures,scudoResults-method}
\alias{downSignatures}
\alias{downSignatures,scudoResults-method}
\alias{groupsAnnotation}
\alias{groupsAnnotation,scudoResults-method}
\alias{consensusUpSignatures}
\alias{consensusUpSignatures,scudoResults-method}
\alias{consensusDownSignatures}
\alias{consensusDownSignatures,scudoResults-method}
\alias{selectedFeatures}
\alias{selectedFeatures,scudoResults-method}
\alias{params}
\alias{params,scudoResults-method}
\alias{show,scudoResults-method}
\title{Class "scudoResults"}
\description{
This is a class that represents the output of the functions -- link to scudo
functions --.
}
\details{
This class provides a structure to represent the results of the scudo
functions. It contains a distance matrix, i.e. a symmetric matrix with
non-negative numeric values.
}
\section{Slots}{

\describe{
\item{\code{distMatrix}}{Object of class "matrix". A symmetric matrix with
non-negative numeric elements.}

\item{\code{upSignatures}}{Object of class "data.frame". A data.frame with the same
colnames as distMatrix, representing the up-regualted features in each
sample.}

\item{\code{downSignatures}}{Object of class "data.frame". A data.frame with the same
colnames as distMatrix, representing the down-regualted features in each
sample.}

\item{\code{groupsAnnotation}}{Object of class "factor". It represents the groups used
for the normalization and the feature selection. It corresponds to the
\code{groupsAnnotation} argument in --link-- and --link--, but unused levels
are dropped.}

\item{\code{consensusUpSignatures}}{Object of class "data.frame". It contains the
consensus signatures of up-regulated features for each group.}

\item{\code{consensusDownSignatures}}{Object of class "data.frame".It contains the
consensus signatures of dowm-regulated features for each group.}

\item{\code{selectedFeatures}}{Object of class "character". A vector of selected
features. If the feature selection was not performed, it contains every
feature present in the input of the scudo functions.}

\item{\code{params}}{Object of class "list". A list of the parameters used to run
--link-- or --link--.}
}}

\section{Methods}{
 \describe{ \item{\code{distMatrix}}{\code{signature(object
  = "scudoResults")}: a method for obtaining the distance matrix.}
  \item{\code{upSignatures}}{\code{signature(object = "scudoResults")}: a
  method for obtaining the signature of up-regualted features in each
  sample.} \item{\code{downSignatures}}{\code{signature(object =
  "scudoResults")}: a method for obtaining the signature of down-regulated
  features in each sample.} \item{\code{groupsAnnotation}}{\code{signature(object
  = "scudoResults")}: a method for obtaining the groups used for
  normalization and feature selection.}
  \item{\code{consensusUpSignatures}}{\code{signature(object =
  "scudoResults")}: a method for obtaining the consensus signatures of
  up-regualted features in each group.}
  \item{\code{consensusDownSignatures}}{\code{signature(object =
  "scudoResults")}: a method for obtaining the consensus signatures of
  down-regulated features in each group.}
  \item{\code{selectedFeatures}}{\code{signature(object = "scudoResults")}: a
  method for obtaining the names of the features seleted. If no feature
  selection was performed, the names of every feature are returned.}
  \item{\code{params}}{\code{signature(object = "scudoResults")}: a method
  for obtaining the parameters that were used to generate the result.} }
}

