% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scudoTest.R
\name{scudoTest}
\alias{scudoTest}
\title{Performs SCUDO analysis on test dataset}
\usage{
scudoTest(trainScudoRes, testExpData, testGroups = NULL, nTop = NULL,
    nBottom = NULL, foldChange = TRUE, groupedFoldChange = FALSE,
    distFun = NULL)
}
\arguments{
\item{trainScudoRes}{an object of class \code{ScudoResult} used as
training model}

\item{testExpData}{either an \code{\link[Biobase]{ExpressionSet}}
or a data.frame or a matrix of gene expression data, with a column for
each sample and a row for each feature}

\item{testGroups}{factor containing group labels for each sample in
\code{testExpData}}

\item{nTop}{number of up-regulated features to include in the signatures}

\item{nBottom}{number of down-regulated features to include in the
signatures}

\item{foldChange}{logical, whether or not to compute fold-changes from
expression data}

\item{groupedFoldChange}{logical, whether or not to take into account the
groups when computing fold-changes. See Details for a description of the
computation of fold-changes}

\item{distFun}{the function used to compute the distance between two
samples. See See \code{\link{scudoTrain}} function Details for the
specification of this function}
}
\value{
Object of class \code{\linkS4class{scudoResults}}.
}
\description{
A function to perform the SCUDO analysis on test data, given an object of
class \code{scudoResults} used as training model.
}
\details{
This function creates an object of class \code{\linkS4class{scudoResults}}

Only the features selected in the trainig step are considered for the
analysis.

Fold-change computation and SCUDO analysis is performed in the same way as
explained in \code{\link{scudoTrain}} function details. In this case, groups
can also not being specified, as well as nTop and nBottom. If the latter are
not specified, same values as the ones in \code{scudoResult} object are used.
}
\examples{
# generate dummy train dataset
exprDataTrain <- data.frame(a = 11:20, b = 16:25,
    c = rev(1:10), d = c(1:2, rev(3:10)))
exprDataTest <- data.frame(e = 1:10, f = 11:20,
    g = rev(11:20), h = c(1:2, rev(3:10)))
rownames(exprDataTrain) <- rownames(exprDataTest) <- letters[11:20]
grpsTrain <- as.factor(c("G1", "G1", "G2", "G2"))
nTop <- 2
nBottom <- 3

# run scudo
res <- scudoTrain(exprDataTrain, grpsTrain, nTop, nBottom,
    foldChange = FALSE, featureSel = FALSE)
show(res)

# run scudoTest
testRes <- scudoTest(res, exprDataTest, foldChange = FALSE)
show(testRes)

}
\seealso{
\code{\link{scudoTrain}}, \code{\link{scudoNetwork}},
\code{\linkS4class{scudoResults}}
}
\author{
Matteo Ciciani \email{matteo.ciciani@gmail.com}
}
